+++ docs/docs/.vuepress/components/FBXLoaderFunction.vue
@@ -1,0 +1,23 @@
+<template>
+  <tv-scene class="scene" @created="created"></tv-scene>
+</template>
+
+<script lang="ts" setup>
+import { FBXLoader } from 'three-vue3'
+
+const created = async (scene, { camera }) => {
+  camera.position.set(0, 1.5, 3)
+
+  // Load model to scene.
+  const model = await FBXLoader('/models/girl.fbx')
+  scene.add(model)
+}
+</script>
+
+<style>
+.scene {
+  margin-top: 10px;
+  width: 100%;
+  height: 300px;
+}
+</style>