+++ src/index.ts
@@ -1,14 +1,24 @@
 import { App } from 'vue'
 import Scene from './components/Scene'
 import GLTFLoader from './components/GLTFLoader'
+import FBXLoader from './components/FBXLoader'
+import OBJLoader from './components/OBJLoader'
+import SkyBox from './utils/SkyBox'
+import Popup from './utils/Popup'
+import ME from './utils/ME'
 
-// components
+// component
 export default {
   install: (app: App) => {
     app.component('tv-scene', Scene)
     app.component('tv-gltf-loader', GLTFLoader)
+    app.component('tv-fbx-loader', FBXLoader)
+    app.component('tv-obj-loader', OBJLoader)
   }
 }
 
-export { SkyBox } from './utils/SkyBox'
-export { Popup } from './utils/Popup'
+// class
+export { SkyBox, Popup, ME }
+
+// function
+export * from './utils/ModelLoader'