+++ src/types/global.d.ts
@@ -1,0 +1,9 @@
+declare global {
+  type Position = [number, number, number]
+  type LoadEvent = {
+    type: string
+    progress: number
+  }
+}
+
+export {}